Decorate Post entity to work with type-graphql as a type
Since the Post entity is a class, type-graphql has some decorator functions that allow us to convert a class into a type by stacking decorators. This means that Post is both an ObjectType and an Entity. Note that Field exposes the entity to the graphql interface. Omitting it keeps the field private.

- date fields type specifically defined
- Int type defined just for practice, but it is not needed as it is inferred
- title field graphql type inferred.