Add login function to user resolver
Login verifies that user exists and that password is valid. Otherwise, it returns a UserResponse which includes a FieldError or the User, with corresponding object being null. For example, if a user is not found, a FieldError is returned and User is null. Otherwise, User is returned and FieldError will be null.