Add error-handling to /register
When user attempts to register, check for errors from graphql endpoint and mark the field with returned error. For example, if the username length validation does not pass, mark the username field and display the error message below the filed. If registration was successful, redirect the user to the "/" route.