Pause useMeQuery hook when on server
Given that index uses SSR and calls NavBar, set the useMeQuery (which fetches the current user) to NOT run when the page is built on the server  to avoid running the query twice. The way we check if on the servder is through the isServer function which checks for window object.