Implement forgotPassword resolver
ForgotPassword generates a key by concatenating a random UUID with a PREFIX, example forget-password:498c45e0-331f-4e5d-9104-b08587235633. The corresponding value is the user.id. The expiration in redis is 3-days. The way this works is that key per above is stored in redis. The token is sent via email to the user only if the user is found in the database. Once the user clicks on the reset password link, the client will strip the token and check redis to get the user id to be able to reset the password.

- constant set for redis forgot password prefix