Invalidate cache for createPost mutation
When creating a post, the cache is invalidated so that the post appears without needing to be refreshed. Given that there can be multiple queries due to the load more, then invalidation loops thorugh all fieldInfos with a fieldName matching "posts". This forces new network requests to be made for all loadMore