Add voteStatus
voteStatus is a graphql field (computed) which is used in order to add logic to the votePanel component in order to limit upvoting or downvoting at client level if the user has already upvoted or downvoted a post. The client-side check is in addition thte server-side check of a user vote on a post.