Add voteStatus to cache fragment update
In order to avoid showing unlimited upvotes of downvotes when a user upvotes of downvotes (at least until refresh when the clients pulls correct values from the server) the voteStatus field is added to the read and write fragments to update cache of posts whenever the voteStatus changes for a post.

This results in post votePanel properly changing between upvote/downvote while still limiting user from "apparently" upvoting or downvoting multiple times until a refresh.