Fwd cookie from nextjs server to graphql server during SSR
When SSR runs, the cookie was not being fwded to the graphql. This resulted in the voteStatus being unavailable since the cookie is what allows the server to know if a user is logged in, getting the userId in that way. Fix by fwding cookie in header for the Urql Client.