Add Edit Post functionality
Edit Post button only shows up if the current user is logged in and is owner of the post in question. When user clicks edit post, teh user is taken to a new form that loads the existing title and text of post and can proceed to edit. When successfully edited, user is taken back to previous page (router.back)

* fix issue in createUrqlClient for cookie value if ctx.req.headers is null (added optional chaining)
* create a new hook, useGetPostFromUrl which also returns the postIdFromRouter in addition to the post query response
* refactor edit/delete button on index to only be visible if current user is post creator
* update /post/[id].tsx (create post page) to use useGetPostFromUrl hook